requirements.txt
intervaltree==3.1.0
transformers==4.48.0
pandas==2.2.3
datasets==3.2.0
torch==2.5.1
evaluate==0.4.3
wandb==0.19.4
seqeval==1.2.2
doxa-cli==0.1.8
python-dotenv==1.0.1